#13 a bit of refactoring
